+++ R/netmeta-package.R
@@ -51,7 +51,10 @@
 #' \item partial order of treatment rankings (\code{\link{netposet}},
 #'   \code{\link{plot.netposet}}) and Hasse diagram
 #'   (\code{\link{hasse}}) according to Carlsen & Bruggemann (2014)
-#'   and Rücker & Schwarzer (2017).
+#'   and Rücker & Schwarzer (2017);
+#' \item VišeKriterijumska Optimizacija I Kompromisno Rešenje (VIKOR)
+#'   multi-criteria decision analysis method (\code{\link{vikor}})
+#'   (Opricovic & Tzeng, 2004).
 #' }
 #' 
 #' Available functions to evaluate network inconsistency:
@@ -190,6 +193,12 @@
 #' Illustrating the assumptions of meta-regression in treatment networks.
 #' Preprint available at \emph{Research Square},
 #' \doi{10.21203/rs.3.rs-8235913/v1}
+#'  
+#' Opricovic S, Tzeng GH (2004):
+#' Compromise solution by MCDM methods: A comparative analysis of VIKOR and
+#' TOPSIS.
+#' \emph{European Journal of Operational Research},
+#' \bold{156}, 445--55
 #' 
 #' Papakonstantinou, T., Nikolakopoulou, A., Rücker, G., Chaimani, A.,
 #' Schwarzer, G., Egger, M., Salanti, G. (2018):
@@ -295,11 +304,11 @@
 #' 
 #' @importFrom grid arrow convertHeight convertWidth convertX convertY drawDetails gpar grid.clip grid.draw grid.lines grid.newpage grid.rect grid.roundrect grid.text grob popViewport pushViewport stringWidth unit viewport
 #' 
-#' @importFrom dplyr %>% filter select rename starts_with relocate last_col mutate if_else bind_rows pull case_when distinct count summarise inner_join
+#' @importFrom dplyr %>% filter select rename starts_with relocate last_col mutate if_else bind_rows pull case_when distinct count summarise inner_join arrange if_any
 #' 
 #' @importFrom stringr str_length str_to_sentence
 #' 
-#' @importFrom tidyr pivot_longer
+#' @importFrom tidyr pivot_longer drop_na
 #' 
 #' @importFrom magrittr %<>%